+++ agora/agora/actions.py
@@ -7,7 +7,7 @@ import json
 from datetime import datetime
 from django.utils import timezone
 from apimas.errors import ValidationError
-from agora.utils import create_eosc_api_json
+from agora.utils import create_eosc_api_json_resource, create_eosc_api_json_provider
 
 EOSC_API_URL = getattr(settings, 'EOSC_API_URL', '')
 OIDC_REFRESH_TOKEN = getattr(settings, 'OIDC_REFRESH_TOKEN', '')
@@ -31,7 +31,7 @@ def get_access_token(oidc_url, refresh_token, client_id ):
     return response.json()['access_token']
 
 def resource_publish_eosc(backend_input, instance, context):
-    eosc_req = create_eosc_api_json(instance)
+    eosc_req = create_eosc_api_json_resource(instance)
     if 'resourceOrganisation' not in eosc_req or eosc_req['resourceOrganisation'] == None or len(eosc_req['resourceOrganisation'].strip()) == 0:
         raise ValidationError('Resource provider has not an eosc_id')
     url = EOSC_API_URL+'resource'
@@ -63,7 +63,7 @@ def resource_publish_eosc(backend_input, instance, context):
 
 
 def resource_update_eosc(backend_input, instance, context):
-    eosc_req = create_eosc_api_json(instance)
+    eosc_req = create_eosc_api_json_resource(instance)
     if 'resourceOrganisation' not in eosc_req or eosc_req['resourceOrganisation'] == None or len(eosc_req['resourceOrganisation'].strip()) == 0:
         raise ValidationError('Resource provider has not an eosc_id')
     url = EOSC_API_URL+'resource'
@@ -76,6 +76,65 @@ def resource_update_eosc(backend_input, instance, context):
         'Content-Type': 'application/json',
     }
     logger.info('EOSC PORTAL API call to PUT resource \
+        with id %s to %s has been made by %s at %s \
+        ' %(id, url, username, datetime.now()))
+    try:
+        response = requests.put(url, headers=headers,json=eosc_req)
+        response.raise_for_status()
+        logger.info('Response status code: %s' %(response.status_code))
+        logger.info('Response json: %s' %(response.json()))
+        instance.eosc_state = "Updated"
+        instance.eosc_updated_at = datetime.now(timezone.utc)
+    except requests.exceptions.RequestException as err:
+        logger.info('Response status code: %s, %s, %s' % (url, err, response.json()))
+        instance.eosc_state = "Error"
+        raise ValidationError("EOSC API: " + response.json()['error'])
+    instance.save()
+    return instance
+
+def provider_publish_eosc(backend_input, instance, context):
+    provider_email = context['auth/user'].email
+    eosc_req = create_eosc_api_json_provider(instance, provider_email)
+    url = EOSC_API_URL+'provider'
+    id  = str(instance.id)
+    username = context['auth/user'].username
+    eosc_token = get_access_token(OIDC_URL, OIDC_REFRESH_TOKEN, OIDC_CLIENT_ID)
+    headers = {
+        'Authorization': eosc_token,
+        'Accept': 'application/json',
+        'Content-Type': 'application/json',
+    }
+    logger.info('EOSC PORTAL API call to POST provider \
+        with id %s to %s has been made by %s at %s \
+        ' %(id, url, username, datetime.now()))
+    try:
+        response = requests.post(url, headers=headers,json=eosc_req)
+        response.raise_for_status()
+        logger.info('Response status code: %s' %(response.status_code))
+        logger.info('Response json: %s' %(response.json()))
+        instance.eosc_state = "Published"
+        instance.eosc_id = response.json()['id']
+        instance.eosc_published_at = datetime.now(timezone.utc)
+    except requests.exceptions.RequestException as err:
+        logger.info('Response status code: %s, %s, %s' % (url, err, response.json()))
+        instance.eosc_state = "Error"
+        raise ValidationError("EOSC API: " +response.json()['error'])
+    instance.save()
+    return instance
+
+def provider_update_eosc(backend_input, instance, context):
+    provider_email = context['auth/user'].email
+    eosc_req = create_eosc_api_json_provider(instance, provider_email)
+    url = EOSC_API_URL+'provider'
+    id  = str(instance.id)
+    username = context['auth/user'].username
+    eosc_token = get_access_token(OIDC_URL, OIDC_REFRESH_TOKEN, OIDC_CLIENT_ID)
+    headers = {
+        'Authorization': eosc_token,
+        'Accept': 'application/json',
+        'Content-Type': 'application/json',
+    }
+    logger.info('EOSC PORTAL API call to PUT provider \
         with id %s to %s has been made by %s at %s \
         ' %(id, url, username, datetime.now()))
     try: