+++ metrics/src/App.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import CssBaseline from '@material-ui/core/CssBaseline';
 
-import logo from './assets/agora.png';
 import Footer from './components/Footer';
 import NavBar from './components/NavBar';
 import TopPage from './components/TopPage';
@@ -11,7 +10,7 @@ const App = () => {
   return (
     <React.Fragment>
       <CssBaseline />
-      <NavBar logo={logo} />
+      <NavBar logo={'logo.png'}/>
       <TopPage />
       <PreviousMonths />
       <Footer />