+++ ui/app/utils/common/eosc-portal.js
@@ -8,9 +8,10 @@ const {
 } = Ember;
 
 const DISABLED = !ENV.APP.eosc_portal.enabled;
-const HIDE_ACTIONS = !ENV.APP.eosc_portal.show_actions;
+const HIDE_ACTIONS_RESOURCE = !ENV.APP.eosc_portal.show_actions_resource;
+const HIDE_ACTIONS_PROVIDER = !ENV.APP.eosc_portal.show_actions_provider;
 
-function prompt(self, method) {
+function promptResource(self, method) {
   const erp_bai_1_name = get(self, 'model.erp_bai_1_name');
   const erp_bai_2_service_organisation = get(self, 'model.erp_bai_2_service_organisation.id');
   const erp_bai_4_webpage = get(self, 'model.erp_bai_4_webpage');
@@ -86,6 +87,64 @@ function prompt(self, method) {
 }
 
 
+function promptProvider(self, method) {
+  const epp_bai_0_id = get(self, 'model.epp_bai_0_id');
+  const epp_bai_1_name = get(self, 'model.epp_bai_1_name');
+  const epp_bai_2_abbreviation = get(self, 'model.epp_bai_2_abbreviation');
+  const epp_bai_3_website = get(self, 'model.epp_bai_3_website');
+  const epp_bai_4_legal_entity = get(self, 'model.epp_bai_4_legal_entity');
+  const epp_loi_1_street_name_and_number = get(self, 'model.epp_loi_1_street_name_and_number');
+  const epp_loi_2_postal_code = get(self, 'model.epp_loi_2_postal_code');
+  const epp_loi_3_city = get(self, 'model.epp_loi_3_city');
+  const epp_loi_5_country_or_territory = get(self, 'model.epp_loi_5_country_or_territory');
+  const epp_mri_1_description = get(self, 'model.epp_mri_1_description');
+  const epp_mri_2_logo = get(self, 'model.epp_mri_2_logo');
+  const main_contact = get(self, 'model.main_contact.id');
+  const public_contact = get(self, 'model.public_contact.id');
+
+  let missing = [];
+  let missing_str = '';
+
+  let noControls = false;
+  let message = `eosc.provider.${method}.message`;
+
+  if (!epp_bai_0_id) { missing.push('epp_bai_0_id') }
+  if (!epp_bai_1_name) { missing.push('epp_bai_1_name') }
+  if (!epp_bai_2_abbreviation) { missing.push('epp_bai_2_abbreviation') }
+  if (!epp_bai_3_website) { missing.push('epp_bai_3_website') }
+  if (!epp_bai_4_legal_entity) { missing.push('epp_bai_4_legal_entity') }
+  if (!epp_loi_1_street_name_and_number) { missing.push('epp_loi_1_street_name_and_number') }
+  if (!epp_loi_2_postal_code) { missing.push('epp_loi_2_postal_code') };
+  if (!epp_loi_3_city) { missing.push('epp_loi_3_city')};
+  if (!epp_loi_5_country_or_territory) { missing.push('epp_loi_5_country_or_territory')};
+  if (!epp_mri_1_description) { missing.push('epp_mri_1_description')};
+  if (!epp_mri_2_logo) { missing.push('epp_mri_2_logo')};
+  if (!main_contact) { missing.push('main_contact')};
+  if (!public_contact) { missing.push('public_contact')};
+
+  const i18n = get(self, 'i18n');
+
+  if (missing.length > 0) {
+    missing.forEach(function(el) {
+      let field = `provider.fields.${el}`;
+      let trans = i18n.t(field);
+      missing_str += `<li>${trans}</li>`;
+    })
+    noControls = true;
+    message = get(self, 'i18n').t(`eosc.provider.${method}.message_missing`, {
+      list_missing: `<ul>${missing_str}</ul>`,
+    });
+  }
+
+  return {
+    ok: `eosc.provider.${method}.ok`,
+    cancel: 'cancel',
+    title: `eosc.provider.${method}.title`,
+    message,
+    noControls,
+  }
+
+}
 const postResourceEOSC = {
   label: 'eosc.resource.post.label',
   icon: 'backup',
@@ -125,7 +184,7 @@ const postResourceEOSC = {
     'model.erp_bai_2_service_organisation',
     function(){
       if ( DISABLED ) { return true }
-      if ( HIDE_ACTIONS ) { return true }
+      if ( HIDE_ACTIONS_RESOURCE ) { return true }
       let role = get(this, 'role');
       let user_org = get(this, 'session.session.authenticated.organisation');
       let state = get(this, 'model.state');
@@ -170,7 +229,7 @@ const postResourceEOSC = {
     'model.erp_mgi_3_terms_of_use',
     'model.erp_aoi_1_order_type.id',
     function(){
-      return prompt(this, 'post');
+      return promptResource(this, 'post');
     }
   ),
 };
@@ -214,7 +273,7 @@ const putResourceEOSC = {
     'model.erp_bai_2_service_organisation',
     function(){
       if ( DISABLED ) { return true }
-      if ( HIDE_ACTIONS ) { return true }
+      if ( HIDE_ACTIONS_RESOURCE ) { return true }
       let resource_org = get(this, 'model.erp_bai_2_service_organisation.id');
       let user_org = get(this, 'session.session.authenticated.organisation');
       let role = get(this, 'role');
@@ -259,7 +318,173 @@ const putResourceEOSC = {
     'model.erp_mgi_3_terms_of_use',
     'model.erp_aoi_1_order_type.id',
     function(){
-      return prompt(this, 'put');
+      return promptResource(this, 'put');
+    }
+  ),
+};
+
+const postProviderEOSC = {
+  label: 'eosc.provider.post.label',
+  icon: 'backup',
+  i18n: Ember.inject.service(),
+	action: function(route, model) {
+		let messages = get(route, 'messageService');
+		let adapter = get(route, 'store').adapterFor('provider');
+		let token = get(route, 'user.auth_token');
+		let url = adapter.buildURL('provider', get(model, 'id'), 'findRecord');
+		return fetch(url + 'publish-eosc/', {
+			method: 'POST',
+			headers: {
+				Accept: 'application/json',
+				Authorization: `Token ${token}`,
+			},
+		})
+		.then(resp => {
+			if (resp.status === 200) {
+				model.reload().then(() => {
+					messages.setSuccess('eosc.provider.post.success');
+				});
+			} else {
+        resp.json().then((data)=> {
+          let err_msg = data && data.details || 'eosc.provider.post.error';
+				  messages.setError(err_msg);
+        })
+			}
+		})
+		.catch(err => {
+			messages.setError('eosc.provider.post.error');
+		});
+  },
+  hidden: computed(
+    'role',
+    'model.state',
+    'model.eosc_id',
+    'model.id',
+    function(){
+      if ( DISABLED ) { return true }
+      if ( HIDE_ACTIONS_PROVIDER ) { return true }
+      let role = get(this, 'role');
+      let user_org = get(this, 'session.session.authenticated.organisation');
+      let state = get(this, 'model.state');
+      let org_id = get(this, 'model.id');
+      let portal_id = get(this, 'model.eosc_id');
+
+      let user_is_provideradmin = role === 'provideradmin';
+      let user_owns_organisation = user_org === org_id;
+      let resource_is_published = state === 'published';
+      let resource_has_eosc_id = portal_id;
+
+      if (
+        user_is_provideradmin &&
+        user_owns_organisation &&
+        resource_is_published &&
+        !resource_has_eosc_id
+      ) {
+        return false;
+      } else {
+        return true;
+      }
+  }),
+  confirm: true,
+  prompt: computed(
+    'model.epp_bai_0_id',
+    'model.epp_bai_1_name',
+    'model.epp_bai_2_abbreviation',
+    'model.epp_bai_3_website',
+    'model.epp_bai_4_legal_entity',
+    'model.epp_loi_1_street_name_and_number',
+    'model.epp_loi_2_postal_code',
+    'model.epp_loi_3_city',
+    'model.epp_loi_5_country_or_territory',
+    'model.epp_mri_1_description',
+    'model.epp_mri_2_logo',
+    'model.main_contact.id',
+    'model.public_contact.id',
+    function(){
+      return promptProvider(this, 'post');
+    }
+  ),
+};
+
+const putProviderEOSC = {
+  label: 'eosc.provider.put.label',
+  icon: 'backup',
+  i18n: Ember.inject.service(),
+	action: function(route, model) {
+		let messages = get(route, 'messageService');
+		let adapter = get(route, 'store').adapterFor('provider');
+		let token = get(route, 'user.auth_token');
+		let url = adapter.buildURL('provider', get(model, 'id'), 'findRecord');
+		return fetch(url + 'update-eosc/', {
+			method: 'POST',
+			headers: {
+				Accept: 'application/json',
+				Authorization: `Token ${token}`,
+			},
+		})
+		.then(resp => {
+			if (resp.status === 200) {
+				model.reload().then(() => {
+					messages.setSuccess('eosc.provider.put.success');
+				});
+			} else {
+        resp.json().then((data)=> {
+          let err_msg = data && data.details || 'eosc.provider.put.error';
+				  messages.setError(err_msg);
+        })
+			}
+		})
+		.catch(err => {
+			messages.setError('eosc.provider.put.error');
+		});
+  },
+  hidden: computed(
+    'role',
+    'model.state',
+    'model.eosc_id',
+    'model.id',
+    function(){
+      if ( DISABLED ) { return true }
+      if ( HIDE_ACTIONS_PROVIDER ) { return true }
+      let role = get(this, 'role');
+      let user_org = get(this, 'session.session.authenticated.organisation');
+      let state = get(this, 'model.state');
+      let org_id = get(this, 'model.id');
+      let portal_id = get(this, 'model.eosc_id');
+
+      let user_is_provideradmin = role === 'provideradmin';
+      let user_owns_organisation = user_org === org_id;
+      let resource_is_published = state === 'published';
+      let resource_has_eosc_id = portal_id;
+
+      if (
+        user_is_provideradmin &&
+        user_owns_organisation &&
+        resource_is_published &&
+        resource_has_eosc_id
+      ) {
+        return false;
+      } else {
+        return true;
+      }
+  }),
+  confirm: true,
+  prompt: computed(
+    'model.epp_bai_0_id',
+    'model.epp_bai_1_name',
+    'model.epp_bai_2_abbreviation',
+    'model.epp_bai_3_website',
+    'model.epp_bai_4_legal_entity',
+    'model.epp_loi_1_street_name_and_number',
+    'model.epp_loi_2_postal_code',
+    'model.epp_loi_3_city',
+    'model.epp_loi_5_country_or_territory',
+    'model.epp_mri_1_description',
+    'model.epp_mri_2_logo',
+    'model.main_contact.id',
+    'model.public_contact.id',
+    function(){
+      return promptProvider(this, 'put');
     }
   ),
 };
@@ -267,4 +492,6 @@ const putResourceEOSC = {
 export {
   postResourceEOSC,
   putResourceEOSC,
+  postProviderEOSC,
+  putProviderEOSC,
 }