+++ CHANGELOG.md
@@ -4,10 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.2](https://github.com/grnet/agora-sp/compare/v1.3.1...v1.3.2) - 2021-06-03
+
+### Added
+- Add Publish provider to EOSC portal actions.
+- Metrics: Enable theme options.
+- Catalog: Filter results by organisation.
+- Metrics: Show historic data
+
 ## [1.3.1](https://github.com/grnet/agora-sp/compare/v1.3.0...v1.3.1) - 2021-05-13
 
 ### Changed
 - Rename accounting application to metrics.
+
 
 ## [1.3.0](https://github.com/grnet/agora-sp/compare/v1.2.0...v1.3.0) - 2021-05-11